Refactor FileUtil.kt and improve documentation.
