+++ app/src/main/java/com/vaultionizer/vaultapp/worker/FileUploadWorker.kt
@@ -11,7 +11,7 @@ import com.vaultionizer.vaultapp.repository.SpaceRepository
 import com.vaultionizer.vaultapp.repository.SyncRequestRepository
 import com.vaultionizer.vaultapp.service.FileExchangeService
 import com.vaultionizer.vaultapp.util.Constants
-import com.vaultionizer.vaultapp.util.writeFileToInternal
+import com.vaultionizer.vaultapp.util.writeFile
 import dagger.assisted.Assisted
 import dagger.assisted.AssistedInject
 import kotlinx.coroutines.Dispatchers
@@ -66,11 +66,7 @@ class FileUploadWorker @AssistedInject constructor(
             }
 
             // Write file to local file system
-            writeFileToInternal(
-                applicationContext,
-                "${request.localFileId}.${Constants.VN_FILE_SUFFIX}",
-                bytes
-            )
+            applicationContext.writeFile(file.localId, bytes)
 
             val vnFile = fileRepository.getFile(request.localFileId)
             vnFile?.remoteId = request.remoteFileId
